@@ -7,7 +7,6 @@
     "alt": "0.18.6",
     "alt-utils": "1.0.0",
     "aviator": "v0.6.1",
-    "babel-eslint": "6.1.2",
     "babel-plugin-transform-object-rest-spread": "6.8.0",
     "babel-preset-es2015": "6.13.2",
     "babel-preset-react": "6.11.1",
@@ -20,19 +19,18 @@
     "debounce": "^1.0.0",
     "deep-equal": "1.0.1",
     "es6-promise-debounce": "^1.0.1",
-    "eslint": "2.10.2",
-    "eslint-plugin-react": "5.1.1",
     "immutable": "~3.8.1",
     "jcampconverter": "^2.6.0",
     "jsdom": "^9.8.0",
     "lodash": "^4.17.0",
     "moment": "^2.14.1",
-    "moment-precise-range-plugin": "^1.2.0",
+    "moment-precise-range-plugin": "^1.2.4",
     "numeral": "^1.5.3",
     "path": "^0.12.7",
     "qrcode.react": "^0.6.1",
     "quagga": "^0.11.5",
     "quill": "1.1.9",
+    "quill-delta": "3.4.3",
     "react": "15.3.2",
     "react-addons-update": "15.3.2",
     "react-barcode": "^1.1.0",
@@ -46,12 +44,11 @@
     "react-input-autosize": "1.1.0",
     "react-notification-system": "^0.2.7",
     "react-qr-reader": "^0.3.1",
-    "react-input-autosize": "1.1.0",
     "react-select": "1.0.0-rc.5",
     "react-stickydiv": "^3.4.19",
     "react-svg-file-zoom-pan": "0.0.6",
     "react-svg-inline": "^1.2.0",
-    "react-ui-tree": "2.5.0",
+    "react-ui-tree": "3.0.0",
     "sha256": "^0.2.0",
     "uuid": "^2.0.2",
     "whatwg-fetch": "1.0.0"
@@ -74,13 +71,17 @@
   },
   "devDependencies": {
     "babel-core": "^6.11.4",
+    "babel-eslint": "6.1.2",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
     "bundle-collapser": "^1.2.1",
     "envify": "^4.0.0",
     "enzyme": "^2.4.1",
-    "eslint": "^3.3.1",
-    "eslint-plugin-react": "^6.1.1",
+    "eslint": "^3.19.0 || ^4.3.0",
+    "eslint-config-airbnb": "^15.1.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-react": "^7.3.0",
     "expect": "^1.20.2",
     "expect-jsx": "^2.6.0",
     "jsdom-global": "2.1.0",
