@@ -5,23 +5,22 @@ import {metPreConv,metPrefSymbols} from './utils/metricPrefix';
 export default class NumeralInputWithUnitsCompo extends Component {
   constructor(props) {
     super(props);
-    let {value, unit,metricPrefix,precision} = props;
+
+    const { value, block, unit, metricPrefix, precision } = props;
     this.state ={
-      unit: unit,
-      value: value,
+      unit,
+      value,
+      block,
       metricPrefix: metricPrefix || "none",
       currentPrecision: precision,
       valueString:  0,
-      showString: false
+      showString: false,
     };
   }
 
   componentWillReceiveProps(nextProps) {
-    let {value, unit} = nextProps;
-    this.setState({
-      unit: unit,
-      value: value,
-    });
+    const { value, unit, block } = nextProps;
+    this.setState({ unit, value, block });
   }
 
   _handleValueChange(value) {
@@ -31,65 +30,73 @@ export default class NumeralInputWithUnitsCompo extends Component {
   }
 
   _handleInputValueChange(event) {
+    const inputField = event.target;
+    const caretPosition = $(inputField).caret();
+    let { value } = inputField;
+    let { valueString } = this.state;
+    const { metricPrefix } = this.state;
+    const lastChar = value[caretPosition - 1] || '';
 
-    let inputField = event.target;
-    let caretPosition = $(inputField).caret();
-    let {value} = inputField;
-    let {metricPrefix,valueString} = this.state;
-    let lastChar =  value[caretPosition-1] || "";
+    if (lastChar !== '' && !lastChar.match(/-|\d|\.|(,)/)) return false;
 
-    if(lastChar != "" && !lastChar.match(/-|\d|\.|(,)/)) return false;
+    const md = lastChar.match(/-|\d/);
+    const mc = lastChar.match(/\.|(,)/);
 
-    let md = lastChar.match(/-|\d/);
-    let mc = lastChar.match(/\.|(,)/);
-
-    if (mc && mc[1]){value = value.slice(0,caretPosition-1)+'.'+value.slice(caretPosition)}
-    else if (value == '0.' || value == '00'){
+    if (mc && mc[1]) {
+      value = `${value.slice(0, caretPosition - 1)}.${value.slice(caretPosition)}`;
+    } else if (value === '0.' || value === '00') {
       value = '0.0';
     }
 
     value = value.replace('--', '');
-    let matchMinus = value.match(/\d+(\-+)\d*/);
-    if(matchMinus && matchMinus[1])
-    value = value.replace(matchMinus[1], '');
+    const matchMinus = value.match(/\d+(\-+)\d*/);
+    if (matchMinus && matchMinus[1]) value = value.replace(matchMinus[1], '');
 
-    if (md||mc){ valueString=value }
+    if (md || mc) valueString = value;
 
-    let val = metPreConv(value, metricPrefix, "none");
+    const val = metPreConv(value, metricPrefix, 'none');
     this.setState({
-        value:  val,
-        showString: true,
-        valueString: valueString
-      }, () => {
-        this._onChangeCallback();
-        $(inputField).caret(caretPosition);
-      }
-    );
-
+      value: val,
+      showString: true,
+      valueString,
+    }, () => {
+      this._onChangeCallback();
+      $(inputField).caret(caretPosition);
+    });
   }
 
   _handleInputValueFocus(){
+    const { value, metricPrefix } = this.state;
+
     this.setState({
       currentPrecision: undefined,
       showString: true,
-      valueString: metPreConv(this.state.value,"none",this.state.metricPrefix) || 0,
-      }, () => {this._onChangeCallback();}
-    );
+      valueString: metPreConv(value, 'none', metricPrefix) || 0,
+    });
   }
 
-  _handleInputValueBlur(){
-     this.setState({
-        currentPrecision: this.props.precision,
-        showString: false
-      }, () => {this._onChangeCallback();}
-  );
+  _handleInputValueBlur() {
+    this.setState({
+      currentPrecision: this.props.precision,
+      showString: false,
+    }, () => this._onChangeCallback());
+  }
+
+  handleInputDoubleClick(event) {
+    if (this.state.block) {
+      this.setState({
+        block: false,
+        value: 0,
+      });
+    }
   }
 
   _onChangeCallback() {
-    if(this.props.onChange) {
+    if (this.props.onChange) {
       this.props.onChange(this.state);
     }
   }
+
   togglePrefix(){
     let {metricPrefixes} = this.props
     let ind = metricPrefixes.indexOf(this.state.metricPrefix)
@@ -101,12 +108,14 @@ export default class NumeralInputWithUnitsCompo extends Component {
     });
   }
 
-// TODO fix css-issue with wrong z-index
   render() {
-    let { bsSize, bsStyle, disabled,label} = this.props;
-    let {unit,showString, value,metricPrefix,currentPrecision,valueString} = this.state;
+    const { bsSize, bsStyle, disabled, label} = this.props;
+    let {
+      unit, showString, value, metricPrefix,
+      currentPrecision, valueString, block,
+    } = this.state;
     let mp = metPrefSymbols[metricPrefix];
-    let val = ()=>{
+    let val = () => {
       const nanOrInfinity = isNaN(value) || !isFinite(value)
       if (!showString && nanOrInfinity){
         return 'n.d.'
@@ -114,29 +123,39 @@ export default class NumeralInputWithUnitsCompo extends Component {
         return  metPreConv(value,"none",metricPrefix).toPrecision(currentPrecision);
       } else {return valueString}
     };
+    let inputDisabled = disabled ? true : block;
 
     let prefixSwitch;
     // BsStyle-s for Input and buttonAfter have differences
     let bsStyleBtnAfter = bsStyle == 'error' ? 'danger' : bsStyle;
     let labelWrap = label ? <ControlLabel>{label}</ControlLabel> : null
     if(unit != 'none') {
-      prefixSwitch = <InputGroup.Button>
+      prefixSwitch = (
+        <InputGroup.Button>
           <Button active
             onClick={() =>{this.togglePrefix()}}
             bsClass='bs-btnTxt--small btn'
             bsStyle={bsStyleBtnAfter}
             bsSize={bsSize}
-            style={{padding:"8px 2px 7px 2px"}} >
-              {mp+unit}
+            style={{padding:"6px 2px 6px 2px"}}
+          >
+            {mp + unit}
           </Button>
         </InputGroup.Button>
+      )
+
       return (
         <div>
           {labelWrap}
-          <InputGroup>
-            <FormControl type='text'
-              bsClass='bs-form--compact form-control'
-              disabled={disabled} bsSize={bsSize} bsStyle={bsStyle}
+          <InputGroup
+            onDoubleClick={(event)=>this.handleInputDoubleClick(event)}
+          >
+            <FormControl
+              type="text"
+              bsClass="bs-form--compact form-control"
+              disabled={inputDisabled}
+              bsSize={bsSize}
+              bsStyle={bsStyle}
               value={val() || ''}
               onChange={(event) => this._handleInputValueChange(event)}
               onFocus={(event) => this._handleInputValueFocus(event)}
@@ -150,14 +169,20 @@ export default class NumeralInputWithUnitsCompo extends Component {
       return(
         <div>
           {labelWrap}
-          <FormControl type='text'
-            bsClass='bs-form--compact form-control'
-            disabled={disabled} bsSize={bsSize} bsStyle={bsStyle}
-            value={val() || ''}
-            onChange={(event) => this._handleInputValueChange(event)}
-            onFocus={(event) => this._handleInputValueFocus(event)}
-            onBlur={(event)=>this._handleInputValueBlur(event)}
-          />
+          <div onDoubleClick={(event)=>this.handleInputDoubleClick(event)}>
+            <FormControl
+              type="text"
+              bsClass="bs-form--compact form-control"
+              disabled={inputDisabled}
+              bsSize={bsSize}
+              bsStyle={bsStyle}
+              value={val() || ''}
+              onChange={(event) => this._handleInputValueChange(event)}
+              onFocus={(event) => this._handleInputValueFocus(event)}
+              onBlur={(event)=>this._handleInputValueBlur(event)}
+              onDoubleClick={(event)=>this.handleInputDoubleClick(event)}
+            />
+          </div>
         </div>
       );
     }
@@ -182,6 +207,7 @@ NumeralInputWithUnitsCompo.defaultProps = {
   value: 0,
   units: [],
   disabled: false,
-  bsSize: "small",
-  bsStyle: "default"
+  block: false,
+  bsSize: 'small',
+  bsStyle: 'default',
 };
