@@ -18,13 +18,13 @@ export default class ModalExport extends React.Component {
         {value: "cano_smiles", text: "canonical smiles", checked: true},
         {value: "sum_formular", text: "sum formula", checked: true},
         {value: "inchistring", text: "inchistring", checked: true},
-        {value: "target_amount_value,target_amount_unit", text: "target amount", checked: false},
+        {value: "target_amount_value, target_amount_unit", text: "target amount", checked: false},
         {value: "created_at", text: "created at", checked: false},
         {value: "updated_at", text: "updated at", checked: false},
         {value: "molfile", text: "molfile", checked: false},
         {value: "purity", text: "purity", checked: false},
         {value: "solvent", text: "solvent", checked: false},
-        {value: "impurities", text: "impurities", checked: false},
+        {value: "molarity_value, molarity_unit", text: "molarity", checked: false},
         {value: "location", text: "location", checked: false},
         {value: "is_top_secret", text: "is top secret?", checked: false},
         {value: "ancestry", text: "ancestry", checked: false},
