@@ -2,9 +2,9 @@ class DetailLevels::Sample
   def base_attributes
     [
       :id, :type, :name, :short_label, :description, :created_at,
-      :target_amount_value, :target_amount_unit, :real_amount_value,
-      :real_amount_unit, :molfile, :purity, :solvent, :impurities, :location,
-      :is_top_secret, :is_restricted, :external_label, :analyses,
+      :target_amount_value, :target_amount_unit, :real_amount_value, :location,
+      :real_amount_unit, :molfile, :solvent, :molarity_value, :molarity_unit,
+      :is_top_secret, :is_restricted, :external_label, :analyses, :purity, 
       :children_count, :parent_id, :imported_readout, :_contains_residues,
       :sample_svg_file, :density, :boiling_point, :melting_point,
       :reaction_description, :container, :pubchem_tag, :xref, :code_log,
@@ -38,8 +38,8 @@ class DetailLevels::Sample
     [
       :description, :container, :analyses, :elemental_compositions,
       :target_amount_value, :target_amount_unit, :real_amount_value,
-      :real_amount_unit, :purity, :solvent, :impurities, :location,
-      :children_count, :parent_id, :imported_readout,
+      :real_amount_unit, :purity, :solvent, :molarity_value, :molarity_unit,
+      :children_count, :parent_id, :imported_readout, :location,
       :boiling_point, :melting_point, :reaction_description,
       :can_update, :can_publish
     ]
