@@ -154,10 +154,11 @@ describe Chemotion::SampleAPI do
             name: 'updated name',
             target_amount_value: 0,
             target_amount_unit: 'g',
+            molarity_value: nil,
+            molarity_unit: 'M',
             description: 'Test Sample',
             purity: 1,
             solvent: '',
-            impurities: '',
             location: '',
             molfile: '',
             is_top_secret: false,
@@ -217,10 +218,11 @@ describe Chemotion::SampleAPI do
             name: 'updated name',
             target_amount_value: 0,
             target_amount_unit: 'g',
+            molarity_value: nil,
+            molarity_unit: 'M',
             description: 'Test Sample',
             purity: 1,
             solvent: '',
-            impurities: '',
             location: '',
             molfile: '',
             is_top_secret: false
@@ -279,11 +281,12 @@ describe Chemotion::SampleAPI do
             name: 'test',
             target_amount_value: 0,
             target_amount_unit: 'g',
+            molarity_value: nil,
+            molarity_unit: 'M',
             external_label: 'test extlabel',
             description: 'Test Sample',
             purity: 1,
             solvent: '',
-            impurities: '',
             location: '',
             density: 0.5,
             boiling_point: 100,
@@ -333,10 +336,11 @@ describe Chemotion::SampleAPI do
             name: 'test',
             target_amount_value: 0,
             target_amount_unit: 'g',
+            molarity_value: nil,
+            molarity_unit: 'M',
             description: 'Test Sample',
             purity: 1,
             solvent: '',
-            impurities: '',
             location: '',
             molfile: '',
             is_top_secret: false
