@@ -38,10 +38,11 @@ const reaction = {
       "target_amount_unit":"g",
       "real_amount_value":null,
       "real_amount_unit":"g",
+      "molarity_value":null,
+      "molarity_unit":"M",
       "molfile":"\n  Ketcher 07081609532D 1   1.00000     0.00000     0\n\n  6  6  0     0  0            999 V2000\n    6.7500   -3.6250    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    7.6160   -4.1250    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    7.6160   -5.1250    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    6.7500   -5.6250    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    5.8840   -5.1250    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    5.8840   -4.1250    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n  1  2  1  0     0  0\n  2  3  2  0     0  0\n  3  4  1  0     0  0\n  4  5  2  0     0  0\n  5  6  1  0     0  0\n  6  1  2  0     0  0\nM  END\n",
       "purity":1,
       "solvent":"",
-      "impurities":"",
       "location":"",
       "is_top_secret":false,
       "is_restricted":false,
@@ -122,10 +123,11 @@ const reaction = {
       "target_amount_unit":"g",
       "real_amount_value":null,
       "real_amount_unit":"g",
+      "molarity_value":null,
+      "molarity_unit":"M",
       "molfile":"\n  Ketcher 07081610292D 1   1.00000     0.00000     0\n\n  1  0  0     0  0            999 V2000\n    8.0500   -4.0500    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\nM  END\n",
       "purity":1,
       "solvent":"",
-      "impurities":"",
       "location":"",
       "is_top_secret":false,
       "is_restricted":false,
@@ -198,10 +200,11 @@ const reaction = {
       "target_amount_unit":"g",
       "real_amount_value":null,
       "real_amount_unit":"g",
+      "molarity_value":null,
+      "molarity_unit":"M",
       "molfile":"\n  Ketcher 07081609562D 1   1.00000     0.00000     0\n\n 12 13  0     0  0            999 V2000\n    1.7321   -2.0000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    0.8660   -1.5000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    0.8660   -0.5000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    1.7321    0.0000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    2.5981   -0.5000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    2.5981   -1.5000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    0.0000    0.0000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   -0.8660   -0.5000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   -1.7321    0.0000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   -1.7321    1.0000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   -0.8660    1.5000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    0.0000    1.0000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n  1  2  1  0     0  0\n  2  3  2  0     0  0\n  3  4  1  0     0  0\n  3  7  1  0     0  0\n  4  5  2  0     0  0\n  5  6  1  0     0  0\n  6  1  2  0     0  0\n  7  8  1  0     0  0\n  8  9  2  0     0  0\n  9 10  1  0     0  0\n 10 11  2  0     0  0\n 11 12  1  0     0  0\n 12  7  2  0     0  0\nM  END\n",
       "purity":1,
       "solvent":"",
-      "impurities":"",
       "location":"",
       "is_top_secret":false,
       "is_restricted":false,
